Refactor payment and ticket handling: Change price type from double to BigDecimal across multiple classes for improved precision and consistency in financial calculations.

+++ src/main/java/ticketsystem/ApplicationLayer/IPaymentService.java
@@ -1,11 +1,13 @@
 package ticketsystem.ApplicationLayer;
 
+import java.math.BigDecimal;
+
 import ticketsystem.DTO.OrderDTO;
 import ticketsystem.DTO.PaymentDetails;
 
 public interface IPaymentService {
 
     boolean connect();
-    boolean pay(double amount, PaymentDetails details);
+    boolean pay(BigDecimal amount, PaymentDetails details);
     
 }
